Support parenthesized types
A type specified like `(OverflowError)` is syntactically not a tuple, as there is no trailing comma, so this is in the AST instead a simple `ast.Name` and will be converted to a `J.Identifier` wrapped in a `J.Parentheses` and wrapped again in a `J.LiteralType`. Possibly the last wrapper should instead be a more generic `ExpressionTypeTree`.